# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"InternalServerError": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"error": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"error",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BancoResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BancoData": {
				"properties": {
					"nombre": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"codigo": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"codigo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BancoWithRelations": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"banco_id": {
						"type": "number",
						"format": "double"
					},
					"nombre": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"codigo": {
						"type": "string"
					}
				},
				"required": [
					"banco_id",
					"nombre",
					"codigo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetBancosResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/BancoWithRelations"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBancoResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBancoData": {
				"properties": {
					"nombre": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"codigo": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "finantek_api",
		"version": "0.0.1",
		"description": "API fir finantek",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Daniel Valdez | hello@daniel-valdez.com | https://daniel-valdez.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/bancos/registrar": {
			"post": {
				"operationId": "RegistrarBanco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/BancoResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "Banco Existe",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "El banco que desea ingresar ya existe en el sistema."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Bancos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BancoData"
							}
						}
					}
				}
			}
		},
		"/api/bancos/ver-todos": {
			"get": {
				"operationId": "VerTodosLosBancos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetBancosResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Bancos"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/bancos/ver/{banco_id}": {
			"get": {
				"operationId": "VerBanco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Bancos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "banco_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/bancos/actualizar/{banco_id}": {
			"put": {
				"operationId": "ActualizarBanco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UpdateBancoResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Bancos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "banco_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBancoData"
							}
						}
					}
				}
			}
		},
		"/api/bancos/eliminar/{banco_id}": {
			"delete": {
				"operationId": "EliminarBanco",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Bancos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "banco_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}